Properly check return value from system() calls
Add proper checks for the system() calls used in pf_pdport.c.  It is not
sufficient to look at the return code to determine success or not, the
return code must be decoded properly using the wait() macros.

Please note, none of the system() calls in pf_pdport.c should really be
there, instead they should be the PNAL layer since they all are just for
Linux systems.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>